requirements.txt
PyYAML==3.10
argparse==1.2.1
requests==2.0.0
wsgiref==0.1.2
